// src/app/models/userLogin.ts
export interface UserLogin {
    userID: number;
    userName: string;
    userPassword: string;
    userType: string;
    userDepartment: string;
    userPicture: string;
    remark: string;
    status: string;
    fullName: string;
    saleManager: string;
    saleTeam: string;
    facManager: string;
    nickname: string;
}